Refactory to Order Table
