+++ src/views/Orders/OrderPage.js
@@ -1,13 +1,13 @@
 import React from "react";
 import { Switch, Route } from "react-router-dom";
-import OrderListPage from "views/Orders/OrderListPage";
+import OrderTablePage from "views/Orders/OrderTablePage";
 // import OrderFormPage from "views/Orders/OrderFormPage";
 
 const OrderPage = () => {
   return (
     <Switch>
       {/* <Route exact path="/orders/:id" component={OrderFormPage} /> */}
-      <Route exact path="/orders" component={OrderListPage} />
+      <Route exact path="/orders" component={OrderTablePage} />
     </Switch>
   );
 };
